refactor(engine): read-side reorg — reads/derivations split, detail and discussion-map renames
- domain/discovery-utils.cjs dissolves into domain/reads.cjs (generic
  manifest/file loads, no phase semantics) and domain/derivations.cjs
  (phase joins, lifecycle, next-phase, cache status); derivations may
  require reads, never the reverse
- lib.cjs: the engine.discovery namespace is replaced by engine.reads +
  engine.derivations, mirroring the split; all five consumers re-pointed
- domain/epic.cjs → domain/epic-detail.cjs, domain/workunit.cjs →
  domain/workunit-detail.cjs — the detail builders stop colliding with
  projections/epic.cjs and projections/workunit.cjs
- projections/discussion.cjs → projections/discussion-map.cjs, mirroring
  the discovery-map ops/projection file pairing
- docs (SKILL.md, library-and-gateway.md, commands.md) updated to the
  new names; test require lines re-pointed and test-discovery-utils.cjs
  renamed to test-reads-derivations.cjs

Zero behaviour change: pure moves/renames — the nine test-gateway-for-*
byte-stability suites pass unmodified.
